# Sped up most_similar and accuracy