I was surprised how long it took to run the "accuracy" function (almost 2 hours with the default settings using the google-trained vectors), so I figured I'd speed it up.  It now runs 54x faster (about 2 minutes).

There were a variety of things slowing it down:
1) it didn't take advantage of the "restricted_vocab" reduction in necessary computation (you don't need to do a dot product of the full array, just the "restricted_vocab" number of vectors
2) you don't need to sort entire arrays (nlogn) just partial arrays (nlogk for most_similar, and for accuracy, just n because you only need the best answer) regardless of whether you're using restricted vocab.

I've left in both the old functions for most_similar and accuracy under their old names, and added most_similar_v2 and accuracy_v2 separately for now for anyone who wants to compare.  I'll clean it up before we merge (there's some more functionality I want to add to most_similar anyway, but one step at a time).  As of now I haven't touched most_similar_cosmul, though the same techniques can speed it up just as much.

most_similar_v2 on its own is about 30% faster than most_similar for the default of topn=10 when using the pre-trained google vectors.  Profiling showed my helper function argTopNSorted_nda is about 10 times faster than argsort, but the performance of most_similar is dominated by matrix multiplication.  

Let me know what you think, and I can move forward with it.  

The other addition I want to make after implementing these speed ups is to add vector negation functionality.  It's not as useful for doing analogies, but it's more useful in document/keyword/concept retrieval than subtracting vectors.  I'll expand on that later but I think you'll like it.  

Cheers

P.S. I forgot to remove some extraneous files before committing, but some of them contain the results of my time tests anyway in case you're interested...