@sebastien-j 2015-05-28
@KCzar , for the analogies, would you mind comparing with pull request #340 ? I made this pull request a few days ago, but didnt complete the unit tests yet.

On my desktop, for the GoogleNews model with default parameters, I get a runtime of ~13.5 seconds. Most of this is spent in reordering the vectors once (from largest to smallest count). If this isn't necessary (do `model.indexed_by_count=True`), it takes ~1.4 seconds.