@KCzar 2015-05-28
Sure, I'll have a look tomorrow.  Is there any chance you can summarize for me where the gains are coming from in your method?  I'm guessing from a cursory overview that you're computing everything all at once from two large matrices...

Looks like we had slightly different goals in mind, though there is certainly overlap.  From what I see you did a pretty heavy rewrite, but didn't touch most_similar and instead removed it entirely from the accuracy function.  I suppose one question is what we want out of the "accuracy" function.

If it is important or useful that "most_similar" be a parameter so that one can experiment with different similarity functions, then my method would be useful, and perhaps worth including in addition to yours.  If the purpose of the "accuracy" function is only to test a trained model, then no question yours is much faster.

Either way, if you haven't changed anything with the most_similar function, then my changes there would be worth adding I think.  I got a modest but useful speedup in that function alone.  

I'll have a closer look tomorrow, but let me know if my analysis is off base.