@sebastien-j 2015-05-28
@KCzar , as I didn't modify the `most_similar` function, I agree that the changes you made may indeed be useful.

My general approach is to use matrix-matrix operations to get arrays of size `num_questions x restrict_vocab` that contain the scores (similarities) for each question. The speed-up comes from doing (mini-)batch computations and from dealing with the two bottlenecks you have already identified.

I replaced `most_similar` with `non_lin`, which transforms the cosine similarity with an arbitrary non-linearity. While reasonably powerful, this approach is a little bit less general.