@KCzar 2015-05-28
Ok - just got a chance to look over the code, and I'm about to clone and test it out.  How important is it to break it up into batches?  I would think the size of the model itself would take up much more memory relative to the additional memory needed to run accuracy.  

At the moment I'm thinking one alternative that would combine our approaches, keep the code simple, and potentially get the best of both worlds would be to modify most_similar so that it can also accept a list of lists for batch processing, and then we can still have accuracy call most_similar.  That may be useful functionality to people even outside of the accuracy function, and I'm sure we could make it run with performance similar to what you're getting.  

What do you think?