@sebastien-j 2015-05-28
For large values of `restrict_vocab`, using smaller mini-batches may be useful. For example, I think a batch of 2000 questions with a 1,000,000 word vocabulary would use about 8 additional GB (and 24GB with 3CosMul).

I haven't thought much about modifying `most_similar`, but it could very well be a fine alternative.