@KCzar 2015-05-28
I just ran your version on my machine.  Certainly fast but I didn't get quite the results you get.  Initial run (unsorted) was about 19.5 seconds, but after it's already sorted, it was still ~6.4 seconds average.  I have a pretty powerful computer so I'm a little surprised our results differ by ~4x.

Just finished running accuracy with the full 3,000,000 word Google vocab, batchsize=2000.  Took 18 minutes, and at peak my memory went up by ~33 extra GB.  That seems quite excessive though.  I would think the code must be copying something unnecessarily.  Numpy dot frequently makes a copy of input arrays before multiplying, but I don't see how that explains the memory usage.  Let me tinker with this a bit.