@KCzar 2015-05-28
Ah - yes I see what's going on.  It is making unnecessary copies, which is also slowing it down a lot.  The batches shouldn't be necessary.  Let me go ahead and see if I can write up a version where most_similar has the option of taking a list of lists or tuples as I think that will be the simplest way of achieving the best of both worlds.  Might need a couple days though.