@KCzar 2015-06-01
Just to update - almost done.  I was able to put in a little time over the weekend and can probably wrap it up tomorrow.  So far so good.  Cheers