@KCzar 2015-06-03
Haven't fully cleaned everything up, but would you mind testing out the version I just pushed?  The functions you'll want are most_similar_v5 and accuracy_v5.

For me, accuracy ran in roughly comparable time to your version, and mostly comparable memory (posts I had found like this were apparently a pack of lies!  ...or just outdated: http://wiki.scipy.org/PerformanceTips).  I think my version may have a brief memory spike that yours does not, however.  Perhaps you could help me there.  

In the version I'm submitting, most_similar_v5 should also be faster than most_similar when topn > 1 (which doesn't come into play during accuracy testing.  The other changes I made to most_similar_v5 are that it can take (and return) lists of lists for bulk processing, though it should be fully backwards compatible.  

I had originally gone the rout of trying to make all of my solutions generalizable so that they didn't necessarily require "indexed_by_count," though ultimately I decided the code was getting too complex, and would be unnecessarily difficult for future contributors to interpret.  

Let me know what you think and we can work from there.