@sebastien-j 2015-06-04
Nice progress! It is still a bit slower, but it's great not to have to change the API.

Using the full 3,000,000 word vocabulary, and `batchsize=200` to avoid memory overuse, my version ran in just over 200 seconds (min 202, max 209 over 3 tries). Yours took from 472 to 474s. In both cases, the total system memory use was about 11 GB.

As you mentioned earlier that you had a reasonably powerful computer, I also find it strange that my tests are much quicker than yours.

In any case, it's nice that your approach is slightly more general, helps simple nearest neighbour search and preserves the user interface. On a different note, it would be preferable if the accuracy logs for each category were displayed as soon as they become available.