@KCzar 2015-06-04
Ah - yeah I haven't considered much about the logging b/c I guess I've had it turned off anyway.  Certainly something we can change though.  

I have no explanation for the speed differences - either between our two computers or between the two methods.  My processor is:

model name      : Intel(R) Xeon(R) CPU E5-1650 v2 @ 3.50GHz
cpu MHz         : 1200.000
cache size      : 12288 KB
cpu cores       : 6

I currently have 64 GB of RAM on this machine too.

I would think the only explanations for how yours is so much faster would be A) maybe you have a lot more cache or B) you are using multithreaded BLAS or something.  

My version took about 1600s last time I ran it, though I was running it with cprofile, which slows it down a bit (shouldn't be THAT much though).  About 2/3 of that time was spent in numpy dot, and the other 1/3 was spent in argmax, so it's further surprising that your method was over 2x faster than mine.  I think I must be doing argmax inefficiently - perhaps you can help me there.  

I'll run each of our versions with the settings you used for comparison and report back, but otherwise I won't have any time until at least the weekend probably.  

What do you think about the general approach though?  I wonder if we should check with Radim regarding what he thinks about my proposed changes to most_similar.