@KCzar 2015-06-04
Ah - yeah, you're just flat outclassing me in terms of hardware, plus multithreaded BLAS.  I'm currently using Anaconda python, so I guess I have to either figure out a way to make that work with multithreaded BLAS (haven't found much in initial google searches) or move off of Anaconda.  

I can certainly simplify it - some parts are unnecessarily convoluted b/c of the evolutionary path it's taken through my initial effort to avoid necessarily having to index_by_count.  I'll see about cleaning it up + commenting over the weekend if I can.

I just ran cprofile versions of each of our methods.  They spend roughly the same amount of time in numpy dot (1211s to 1154s), but mine is spending way too much time in argmax, as I suspected ( 352s to 26s), even though argmax is being called about the same number of times.  The reason for that undoubtedly has to do with the order (c contiguous vs f contiguous) of the array that is being input into argmax.  I will have to rethink my code to make sure I get an array in the correct format before running argmax on it.  

Overall, your version ran in 1197s compared to mine, which ran in 1567s with no vocab restrictions and batchsize = 200.

The argmax issue would also explain why you're getting > 2x speedup, because given that you're using multithreaded BLAS, numpy dot is probably a much lower proportion of your computational expense.  

I'll get back to you on or after the weekend.